Abstract processor code to protocol, regex base class.
